requirements.txt
pandas>=1.3.0
numpy>=1.21.0
yfinance>=0.2.0
requests>=2.26.0
pytz>=2021.3
python-dotenv>=0.19.0
schedule>=1.1.0
